feat: add runs-on magic cache
In v2.6.3 of runs-on they introduced a new feature named "Magic Cache". It redirects the cache calls for actions like `actions/cache` and `docker/build-push-action` to a S3 bucket, which is faster for our private runners.
If the action is run on a official runner, it just does nothing, so it is safe to be called on both runner types.